chore: move StringArray to schedule
